Fix #9
Actually two bugs: one is peripherals getting turned on and off again
quickly - this can cause the laser module to draw too much power and shut
down the regulator. In combination with BT module, this turns off the error
beep as it is powered via the BT module. Display also does not work. This is
fixed with a simple firmware update, which will be released imminently.

Second bug appears to be in the bootloader code. Port C9 is the button, and
currently only responds to a falling edge during sleep. If the button is pressed
on entry to sleep, the processor believes it is permanently held down,
and will not respond to falling edges. Fix is to make sensitive to changes in
either direction.

To correct the second bug "in the field" will need to reflash the bootloader
code, which will require either disassembly of the SAP and reprogramming
with a PicKit or ICD3 or SNAP, or (more simply) a bootloader firmware update
whizzy thing. This will be my next focus of code.

Signed-off-by: Phil Underwood <beardydoc@gmail.com>